WIP: moves chart functions into separate module; somehow git line parsing is not working as expected, and some newly added functions were mixed in with these changes
